Sonar and Uptime counter fixes

+++ CatoMat/AppServer.h
@@ -147,10 +147,10 @@ private:
 		else
 			sendMsg(OUTBOX_MODE, "manual");	
 
-		int totalSeconds = millis() / 1000;
-		int seconds = totalSeconds % 60;
-		int minutes = (totalSeconds % (60 * 60)) / 60;
-		int hours = totalSeconds / (60 * 60);
+		unsigned long totalSeconds = millis() / 1000;
+		unsigned long seconds = totalSeconds % 60;
+		unsigned long minutes = (totalSeconds % (60 * 60)) / 60;
+		unsigned long hours = totalSeconds / (60 * 60);
 
 		String uptime = String(hours) + ":";
 		if (minutes < 10)
